Rename Resucito.Blazor => Dictionary
