+++ Dictionary.Shared/Dictionary.Shared.csproj
