+++ Dictionary.sln
@@ -9,9 +9,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Resucito.Blazor", "Resucito.Blazor\Resucito.Blazor.csproj", "{E523BB85-D72F-4FC4-AE52-65986613E392}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dictionary", "Dictionary\Dictionary.csproj", "{E523BB85-D72F-4FC4-AE52-65986613E392}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Resucito.Blazor.Shared", "Resucito.Blazor.Shared\Resucito.Blazor.Shared.csproj", "{3BD5FDE8-9013-4A95-8C89-A4392CA34B16}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Dictionary.Shared", "Dictionary.Shared\Dictionary.Shared.csproj", "{3BD5FDE8-9013-4A95-8C89-A4392CA34B16}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
