+++ Dictionary/Dictionary.csproj
@@ -13,7 +13,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\Resucito.Blazor.Shared\Resucito.Blazor.Shared.csproj" />
+		<ProjectReference Include="..\Dictionary.Shared\Dictionary.Shared.csproj" />
 	</ItemGroup>
 
 </Project>
