+++ Dictionary/Program.cs
@@ -8,7 +8,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace Resucito.Blazor
+namespace Dictionary
 {
     public class Program
     {
