+++ Dictionary/_Imports.razor
@@ -5,5 +5,5 @@
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using Resucito.Blazor
-@using Resucito.Blazor.Shared
+@using Dictionary
+@using Dictionary.Shared
