+++ Dictionary/wwwroot/sample-data/weather.json
